backend/pyproject.toml
[project]
name = "knodex-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "uvicorn[standard]==0.34.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "pydantic[email]==2.10.3",
    "pydantic-settings==2.7.0",
    "python-multipart==0.0.20",
    "aiofiles==24.1.0",
    "boto3==1.35.99",
    "pymupdf==1.25.1",
    "tiktoken==0.8.0",
    "arq==0.26.3",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "httpx==0.28.1",
    "firebase-admin>=6.5.0",
]

[dependency-groups]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "httpx==0.28.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
testpaths = ["tests"]
